Wrap up initial development.
When I last worked on this plugin I had sketched out it's basic structure
but it was all just un-executed code spikes to determine to pieces.

Since that time Daniel Honig had started on adjusting my basic structure
to use a 3rd party Smarty Streets gem for the API integration. Although the
API interaction was trivial using a gem was a better decision as it gives
access to richer data should this plugin continue to be enhanced.

This commit continues that work by transforming a sketched out basic structure
into working tested source code. Changes made:

* I got rid of my old module that I had planned to use to implement the API
  interaction (app/models/concern/smarty_streets_address_verification.rb).
* I got rid of what I believe was just some example tested code that Daniel
  had imported (spec/address_verifier_spec.rb and lib/AddressVerifier.rb).
* I got rid of the module that Daniel was working on that I believe was trying
  to merge what I was trying to do with my module and what the example code
  was trying to do. I don't think this module is necessary. The only reason
  I had created a module was to abstract the API interaction. Since the 3rd
  party gem is doing that for us I believe we can just put everything in the
  decorator and keep the code simpler.
* I moved the environment configuration out of the check and into the loading
  process. This way if an application doesn't have the environment variables
  set it will error earlier.

There are a few other minor bits of cleanup:

* When Daniel imported my rough draft I had accidently left a file created
  by my editor (.zedstate).
* I included dotenv into the test process to ease setting the environment
  variables during testing.
* I updated the README file to be more current.