Remove unnecessary testing support.
The base of this extension was generated by Spree's generator. It provides
a boatload of testing support to do testing at all levels of the stack.

Since this extension just decorates at the model level much of that support
is unncessary and slows down the testing process. This commit removes that
unnecessary support leaving just the testing facilities we are using.