Add test to ensure connectivity or configuration errors bubble up.
About to call this addon completed when I noticed rcov indicating I failed
to write a test for the case where a general connectivity or configuration
error has occurred.

We are just wanting to bubble that error up so the ops team can detect the
issue and resolve it. This tests verifies this so we know we are not swallowing
the error.