Add VCR
When I originally wrote this I didn't know about VCR and I wasn't a
bit fan of manually mocking due to drift between the real system and
a developers idea of of how that system worked. VCR solves that
problem by mocking but based on real interactions that can be
peridically cleared to re-validate against a real system.

This has become even more critical as I think when I first developed
this plugin the free account came with 10,000 lookups and this has been
reduced to 500 which can be zapped quickly.