Add support for client-side auto-complete and verification.
The primary line of defense is server-side validation. But for a better
user experience we can also use SmartyStreet's jQuery plugin to provide both
auto-complete and validation.

Due to a bug in SmartyStreets plugin this is more complicated than it needs
to be. I have been told by SmartyStreets that it will be fixed in the coming
weeks so we should be able to simplify this significantly.