Remove auto-verification.
Per the Smartystreets documentation:

    In order to stay fast, address suggestions as you type are not
    guaranteed to be valid, nor do they come with ZIP Codes.
    Addresses must still go through the verification process.
    Autocomplete does not use any lookups against your account;
    lookups are only deducted from your balance when when the address
    is submitted for verification. Autocomplete suggestions are also
    not guaranteed to be in standardized format or cased properly.

For a real address this isn't horrible as generally real addresses will
validate. But for demo purposes this can look kind of bad because people
often pick the first address which may not be a valid address. To make
it demo better (and still perform fine as it is validated on submission)
we are removing the autoVerify as soon as enough of the address is known.

The only downside to this is that the user may end up filling in the zip
code (since those dont come with auto-complete) while had it auto-verified
it would have filled in the zip code during verification.