Disable `select2`.
Maybe we could figure out a way to make the smarty streets jQuery plugin
automatically populate a select2 field (hook into their events), but really
I find select2 annoying anyway. Sure it provides a better UI than a normal
browser. But it also deviates from normal browsers and this is just one
example where that deviation causes problems. If you are going to override
stock browser behavior make sure it is a complete drop-in.

There really aren't that many states so just switching to to a plain normal
select list.