Make SmartyStreets initlization compatible with Turbolinks.
Resolves issue where smarty streets works if you do a full page load (i.e. typical
when developing) but not when doing a turbolink-enabled page load (i.e. typical for
real world users).

When intializing smarty streets we hooked into the standard jQuery `ready` event.
I had assumed the jQuery/Turbolinks[1] compatiblity library was loaded but this
bug made it apparent this wasn't the case. Don't want to worry about adding that
library introducing new issues so just manually tied to both the jQuery ready event
and the Turbolinks `page:load` event.

1. https://github.com/kossnocorp/jquery.turbolinks