Use more reliable method of fixing up state select after new states loaded.
Previously I waited 500ms after we asked for new states to "fix up" the select
so that the `select2` plugin is disabled. In my earlier testing this was pretty
reliable but later testing has shown it to be fairly unreliable. If the states
take longer than 500ms to load the select doesn't get fixed up. This prevents
smarty streets from being able to interact with the state which also prevents the
validation from working correctly.

The new implementation now appends to the already existing `done` callback so that
we always only execute AFTER the states have loaded.