Rewrote most of the JS integration.
There were several corner cases that were not well addressed with the
old code and attempting to tweak it just resulted into a wack-a-mole
situtation.

Main Integration (address-verification.js.coffee)
-------------------------------------------------

The main integration has been simplified to do the following things:

* When the page is loaded the plugin is intiailized
* There is a public method to disable the verification on an address
  given any field from that address.
* There is a public method to re-enable the verification on an address
  given any field from that address.

The `smarty_disable` function can be just for a moment (for example
when shipping is hidden or show) or it can be made permanent if the
optional second argument is true. The reason to permanently disable is
that there are certain situtations where the plugin reaches a state
where it fails to work and there is no way to get it working again.
Tather than it end up in that broken state we actively disable it
proactively.

The primary use case of this permanent disable is when the country
changes. The plugin cannot deal with the state options being swapped
out. When this happens it will always blank out the state causing the
address save to always fail. The primary situtation this would happen
is when the address is Canada and we want to edit it to be US. Switching
to US previously re-enabled the plugin. Now we leave it disabled even
though we should be able to validate the address. Server-side validation
is still in place.

This primary use case is account for in this commit. The main integration
checks for non-US address during load.

Overrides
---------

To make the plugin work better with Spree there is a bunch of Spree
functionality we tweak. It is as follows:

* If we think the country got changed (due to the states being updated)
  we permanently disable the validation (even if the country became US)
  due to the previously mentioned plugin bug with state options. The
  states are loaded in two different ways (one for the frontend and one
  for the backend). This decorates both state updates.
* The select2 functionality is disabled on the state dropdown. The plugin
  cannot deal well with it. It also hooks into the state updates as Spree
  will try to re-initialize select2.
* If the "Use Billing" option is selected we tweak Spree so that it will
  enable/disable the validation (not permanently) so we don't get validation
  errors on the hidden field which may or may not be valid but it doesn't
  matter.
* When a customer is selected on an order in the backend we send a change
  event as the plugin doesn't actively monitor the fields but instead listens
  for the change event which spree does not send by default.